Cache-bust all jsonp fetches
I noticed on iOS that the back button would sometimes cause the data to
be stale. I believe this could be some internal caching strategy used by
safari. I'm hoping that a true cache bust will help. The old cache-bust
was a way for me to try to force caching to happen, but that was before
I added the stale-fetch-update strategy.