Allow admins (me) to hide competitions directly
Sometimes there's a competition in the list that we don't want to be
there. Removing it is a pain because it will get synced back in again if
it's still in the list from golfbox. Better to have a flag set to hide
it, and allow me to do it straight from the website.